OneClassSVM: Move to outlier_detection module
